refactor to shared handle convo response fxn
